+++ packages/rxget/lib/src/get_state_manager/get_state_manager.dart
@@ -8,4 +8,5 @@ export 'src/rx_flutter/rx_ticket_provider_mixin.dart';
 export 'src/simple/get_controllers.dart';
 export 'src/simple/get_state.dart';
 export 'src/simple/get_view.dart';
+export 'src/simple/list_notifier.dart' show RxLifecycleDebug;
 export 'src/simple/simple_builder.dart';