+++ packages/rxget/lib/src/get_state_manager/src/rx_flutter/rx_notifier.dart
@@ -1,5 +1,6 @@
 import 'dart:async';
 
+import 'package:equatable/equatable.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import '../../../../rxget.dart';
@@ -278,7 +279,7 @@ extension StateExt<T> on StateMixin<T> {
 
 typedef NotifierBuilder<T> = Widget Function(T state);
 
-abstract class GetStatus<T> with Equality {
+abstract class GetStatus<T> with Equatable {
   const GetStatus();
 
   factory GetStatus.loading() => LoadingStatus<T>();