+++ packages/rxget/lib/src/get_utils/get_utils.dart
@@ -1,2 +1,1 @@
-export 'src/equality/equality.dart';
 export 'src/extensions/export.dart';